Rename mint-coins-candidates.js to mint-coins-to-candidates.js
