+++ scripts/mint-coins-to-candidates.js
